Update use case description
The three requests received from @mlagally for updating this document were modified as follows.
o 'Security consideration' section was modified. But, this document deals with an use case, so I think that it is undesirable to include specific technical details for sercurity consideration.
o 'Expected devices' setion and 'Description' section were modified to reflect the 2nd update suggestion from @mlagally.
o 'gap' section was modified to reflect the 3rd updata suggestioin from @mlagally.

And, I received on mote request from @egekorkan to add a section decsribing new Requirement including protocol, content type, platform, and authentication in relation to pull requrest #193. However, this request is currenlty under discussion to improve the WoT use case template. So, I will review it again when the discussion on this is complete.