Rename WoT-UseCases-Template.yml to WoT-UseCases-Template.yml.bak
There are a old version of a Use Case Template "WoT-UseCases-Template.yml". If it is there, submitters could make a mistake when creating a Use Case Issue using the Use Case Template.  As a result, that file no longer appears.

I renamed the file instead of removing it because that file could be reused.